+++ contracts/axelar-gas-service/src/lib.rs
@@ -10,6 +10,9 @@ mod interface;
 #[cfg(all(target_family = "wasm", feature = "testutils"))]
 compile_error!("'testutils' feature is not supported on 'wasm' target");
 
+#[cfg(any(test, feature = "testutils"))]
+pub mod testutils;
+
 cfg_if::cfg_if! {
     if #[cfg(all(feature = "library", not(feature = "testutils")))] {
         pub use interface::{AxelarGasServiceClient, AxelarGasServiceInterface};
