+++ contracts/axelar-gas-service/src/testutils.rs
@@ -1,0 +1,30 @@
+use soroban_sdk::testutils::Address as _;
+use soroban_sdk::token::StellarAssetClient;
+use soroban_sdk::{Address, Env};
+use stellar_axelar_std::types::Token;
+
+use crate::{AxelarGasService, AxelarGasServiceClient};
+
+pub fn setup_gas_service<'a>(env: &Env) -> AxelarGasServiceClient<'a> {
+    let owner: Address = Address::generate(env);
+    let gas_collector: Address = Address::generate(env);
+    let gas_service_id = env.register(AxelarGasService, (&owner, &gas_collector));
+    let gas_service_client = AxelarGasServiceClient::new(env, &gas_service_id);
+
+    gas_service_client
+}
+
+pub fn setup_gas_token(env: &Env, sender: &Address) -> Token {
+    let asset = &env.register_stellar_asset_contract_v2(Address::generate(env));
+    let gas_amount: i128 = 1;
+    let gas_token = Token {
+        address: asset.address(),
+        amount: gas_amount,
+    };
+
+    StellarAssetClient::new(env, &asset.address())
+        .mock_all_auths()
+        .mint(sender, &gas_amount);
+
+    gas_token
+}
