+++ contracts/interchain-token-service/tests/deploy_remote_canonical_token.rs
@@ -1,14 +1,17 @@
+mod utils;
+
 use soroban_sdk::testutils::{Address as _, AuthorizedFunction, AuthorizedInvocation};
 use soroban_sdk::token::{self, StellarAssetClient};
 use soroban_sdk::{Address, Bytes, BytesN, IntoVal, String, Symbol};
 use soroban_token_sdk::metadata::TokenMetadata;
+use stellar_axelar_gas_service::testutils::setup_gas_token;
 use stellar_axelar_std::address::AddressExt;
 use stellar_axelar_std::{auth_invocation, events, mock_auth};
 use stellar_interchain_token_service::event::InterchainTokenDeploymentStartedEvent;
-use stellar_interchain_token_service::testutils::{setup_env, setup_gas_token, TokenMetadataExt};
 use stellar_interchain_token_service::types::{
     DeployInterchainToken, HubMessage, Message, TokenManagerType,
 };
+use utils::{setup_env, TokenMetadataExt};
 
 #[test]
 fn deploy_remote_canonical_token_succeeds() {
