+++ contracts/interchain-token-service/tests/interchain_transfer.rs
@@ -1,12 +1,14 @@
+mod utils;
+
 use soroban_sdk::testutils::Address as _;
 use soroban_sdk::{Address, Bytes, BytesN, Env, String};
+use stellar_axelar_gas_service::testutils::setup_gas_token;
 use stellar_axelar_std::traits::BytesExt;
 use stellar_axelar_std::{assert_contract_err, events};
 use stellar_interchain_token_service::error::ContractError;
 use stellar_interchain_token_service::event::InterchainTransferSentEvent;
-use stellar_interchain_token_service::testutils::{
-    register_chains, setup_env, setup_gas_token, setup_its_token,
-};
+use stellar_interchain_token_service::testutils::setup_its_token;
+use utils::setup_env;
 
 fn dummy_transfer_params(env: &Env) -> (String, Bytes, Option<Bytes>) {
     let destination_chain = String::from_str(env, "ethereum");
@@ -70,7 +72,9 @@ fn interchain_transfer_send_fails_when_paused() {
 #[should_panic(expected = "burn, Error(Contract, #9)")]
 fn interchain_transfer_send_fails_on_insufficient_balance() {
     let (env, client, _, _, _) = setup_env();
-    register_chains(&env, &client);
+    client
+        .mock_all_auths()
+        .set_trusted_chain(&client.its_hub_chain_name());
 
     let sender: Address = Address::generate(&env);
     let gas_token = setup_gas_token(&env, &sender);
@@ -93,7 +97,9 @@ fn interchain_transfer_send_fails_on_insufficient_balance() {
 #[test]
 fn interchain_transfer_fails_on_zero_amount() {
     let (env, client, _, _, _) = setup_env();
-    register_chains(&env, &client);
+    client
+        .mock_all_auths()
+        .set_trusted_chain(&client.its_hub_chain_name());
 
     let sender: Address = Address::generate(&env);
     let gas_token = setup_gas_token(&env, &sender);
@@ -119,7 +125,9 @@ fn interchain_transfer_fails_on_zero_amount() {
 #[test]
 fn interchain_transfer_fails_on_empty_destination_address() {
     let (env, client, _, _, _) = setup_env();
-    register_chains(&env, &client);
+    client
+        .mock_all_auths()
+        .set_trusted_chain(&client.its_hub_chain_name());
 
     let sender: Address = Address::generate(&env);
     let gas_token = setup_gas_token(&env, &sender);
@@ -146,7 +154,9 @@ fn interchain_transfer_fails_on_empty_destination_address() {
 #[test]
 fn interchain_transfer_fails_on_empty_data() {
     let (env, client, _, _, _) = setup_env();
-    register_chains(&env, &client);
+    client
+        .mock_all_auths()
+        .set_trusted_chain(&client.its_hub_chain_name());
 
     let sender: Address = Address::generate(&env);
     let gas_token = setup_gas_token(&env, &sender);
@@ -173,7 +183,9 @@ fn interchain_transfer_fails_on_empty_data() {
 #[test]
 fn interchain_transfer_fails_with_invalid_token() {
     let (env, client, _, _, _) = setup_env();
-    register_chains(&env, &client);
+    client
+        .mock_all_auths()
+        .set_trusted_chain(&client.its_hub_chain_name());
 
     let sender: Address = Address::generate(&env);
     let gas_token = setup_gas_token(&env, &sender);
