+++ contracts/interchain-token-service/tests/message_routing.rs
@@ -1,12 +1,14 @@
+mod utils;
+
 use soroban_sdk::testutils::Address as _;
 use soroban_sdk::{Address, Bytes, String};
+use stellar_axelar_gas_service::testutils::setup_gas_token;
 use stellar_axelar_std::address::AddressExt;
 use stellar_axelar_std::assert_contract_err;
 use stellar_axelar_std::traits::BytesExt;
 use stellar_interchain_token_service::error::ContractError;
-use stellar_interchain_token_service::testutils::{
-    register_chains, setup_env, setup_gas_token, setup_its_token,
-};
+use stellar_interchain_token_service::testutils::setup_its_token;
+use utils::setup_env;
 
 #[test]
 fn send_directly_to_hub_chain_fails() {
@@ -34,7 +36,9 @@ fn send_directly_to_hub_chain_fails() {
 #[test]
 fn send_to_untrusted_chain_fails() {
     let (env, client, _gateway_client, _, _) = setup_env();
-    register_chains(&env, &client);
+    client
+        .mock_all_auths()
+        .set_trusted_chain(&client.its_hub_chain_name());
 
     let sender: Address = Address::generate(&env);
     let amount = 1;
