+++ contracts/interchain-token-service/tests/register_canonical_token.rs
@@ -1,3 +1,5 @@
+mod utils;
+
 use soroban_sdk::testutils::Address as _;
 use soroban_sdk::xdr::ToXdr;
 use soroban_sdk::{Address, BytesN};
@@ -5,8 +7,8 @@ use stellar_axelar_std::address::AddressExt;
 use stellar_axelar_std::{assert_contract_err, events};
 use stellar_interchain_token_service::error::ContractError;
 use stellar_interchain_token_service::event::InterchainTokenIdClaimedEvent;
-use stellar_interchain_token_service::testutils::setup_env;
 use stellar_interchain_token_service::types::TokenManagerType;
+use utils::setup_env;
 
 #[test]
 fn register_canonical_token_succeeds() {
